WHIAD-24675 read ad price as float instead of integer

+++ src/SPT-Pulse/816_SPT_map_search_result_item.js
@@ -112,7 +112,7 @@ b.make_classified_ad_iad = function(ad) {
         publisher: b.resolve_publisher(ad),
         publisherType: ad.publisherType ? ad.publisherType.toLowerCase() : ad.publisherType,
         location: ad_location,
-        price: parseInt(ad.price),
+        price: parseFloat(ad.price),
     };
     result["@id"] = sdrn;
     result["@type"] = "ClassifiedAd";
